# apps/a13n-harness-ui/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harness UI",
    "version": "1"
  },
  "paths": {
...
    "/api/operations/{receipt_id}": {
      "get": {
        "summary": "Operation",
        "operationId": "operation_api_operations__receipt_id__get",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootOperationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContinuationSelectionView": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "selected",
              "not_available",
              "failed"
            ],
            "title": "Status"
          },
          "continuation_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9a-f]{64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Continuation Id"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ContinuationSelectionView"
      },
...
      "EnvironmentOutcomeView": {
        "properties": {
          "unchanged": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Unchanged"
          },
          "published": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Published"
          },
          "failed": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Failed"
          },
          "cleanup_failures": {
            "items": {
              "$ref": "#/components/schemas/FailureView"
            },
            "type": "array",
            "title": "Cleanup Failures",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "unchanged",
          "published",
          "failed"
        ],
        "title": "EnvironmentOutcomeView"
      },
...
      "FailureView": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Code"
          },
          "message": {
            "type": "string",
            "maxLength": 32768,
            "minLength": 1,
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "retry_hint": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Hint"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "FailureView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JsonValue": {},
...
      "RootExecutionView": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed",
              "cancelled",
              "suspended"
            ],
            "title": "Status"
          },
          "output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "output_omitted": {
            "type": "boolean",
            "title": "Output Omitted",
            "default": false
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureView"
              },
              {
                "type": "null"
              }
            ]
          },
          "usage": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "RootExecutionView"
      },
      "RootOperationStatus": {
        "type": "string",
        "enum": [
          "preparing",
          "running",
          "completed",
          "suspended",
          "failed",
          "cancelled"
        ],
        "title": "RootOperationStatus"
      },
      "RootOperationView": {
        "properties": {
          "receipt": {
            "$ref": "#/components/schemas/RootRunReceipt"
          },
          "status": {
            "$ref": "#/components/schemas/RootOperationStatus"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RootRunOutcomeView"
              },
              {
                "type": "null"
              }
            ]
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureView"
              },
              {
                "type": "null"
              }
            ]
          },
          "available_actions": {
            "items": {
              "type": "string",
              "enum": [
                "wait",
                "steer",
                "cancel"
              ]
            },
            "type": "array",
            "title": "Available Actions",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "receipt",
          "status"
        ],
        "title": "RootOperationView"
      },
      "RootRunOutcomeView": {
        "properties": {
          "execution": {
            "$ref": "#/components/schemas/RootExecutionView"
          },
          "continuation": {
            "$ref": "#/components/schemas/ContinuationSelectionView"
          },
          "environment": {
            "$ref": "#/components/schemas/EnvironmentOutcomeView"
          },
          "composition_id": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Composition Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "execution",
          "continuation",
          "environment",
          "composition_id"
        ],
        "title": "RootRunOutcomeView"
      },
      "RootRunReceipt": {
        "properties": {
          "receipt_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Receipt Id"
          },
          "thread_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Thread Id"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "receipt_id",
          "thread_id",
          "submitted_at"
        ],
        "title": "RootRunReceipt"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}